server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
web3==6.0.0
python-dotenv==1.0.0
